NP-1668: split examples in own module
