+++ event-handler-template-examples/src/main/java/no.unit.nva.events.examples/DestinationsEventConsumer.java
