+++ event-handler-template-examples/src/main/java/no.unit.nva.events.examples/EventProducer.java
