+++ event-handler-template-examples/src/test/java/no/unit/nva/events/examples/EventConsumerTest.java
