+++ event-handler-template-examples/src/test/resources/notJsonFile.txt
