+++ event-handler-template-examples/src/test/resources/validAwsEventBridgeEvent.json
