+++ event-handler-template-examples/src/test/resources/validEventBridgeEvent.json
@@ -1,0 +1,19 @@
+{
+  "version": "0",
+  "id": "c2c057c1-67fe-1208-fc21-0401596631ba",
+  "detail-type": "The detail type",
+  "source": "lambda",
+  "account": "884807050265",
+  "time": "2020-10-09T09:34:33Z",
+  "region": "eu-west-1",
+  "resources": [
+    "arn:aws:events:eu-west-1:884807050265:event-bus/orestis-event-bus",
+    "arn:aws:lambda:eu-west-1:884807050265:function:orestis-stack-nva-datacite-mds-EventProducer-8IAE7WYCNTPX:$LATEST"
+  ],
+  "detail": {
+    "publicationId": "https://localhost/fromOutput",
+    "existingDoi": "http://somedoi.org",
+    "xml": "Somexml",
+    "type": "MyType"
+  }
+}
