+++ event-handler-template/src/main/java/no/unit/nva/events/handlers/DestinationsEventBridgeEventHandler.java
