+++ event-handler-template/src/main/java/no/unit/nva/events/models/AwsEventBridgeEvent.java
