+++ event-handler-template/src/main/java/no/unit/nva/events/models/JsonSerializable.java
