+++ event-handler-template/src/main/resources/log4j2.xml
