+++ event-handler-template/src/test/java/no/unit/nva/events/handlers/WithType.java
