+++ event-handler-template/src/test/resources/notJsonFile.txt
@@ -1,0 +1,1 @@
+{hello}
