+++ event-handler-template/src/test/resources/validEventBridgeEvent.json
