+++ settings.gradle
@@ -2,4 +2,5 @@ rootProject.name = 'nva-datacite-mds'
 include 'assign-doi-proxy-function'
 include 'dynamodb-stream-eventbridge-fanout'
 include 'assign-doi-datacite'
-include 'dynamo-event-to-datacite-request'
+include 'event-handler-template'
+include 'event-handler-template-examples'
